VTKHeaders warning
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@4342 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ ShadingArrayController.m
@@ -73,7 +73,7 @@
 	_enableEditing = enable;
 }
 
-- (BOOL)setSelectionIndex:(unsigned int)index
+- (BOOL)setSelectionIndex:(NSUInteger)index
 {	
 	BOOL v = [super setSelectionIndex:(unsigned int)index];
 	[winController applyShading: self];
